requirements.txt
crewai>=0.28.0
crewai-tools>=0.12.0
openai>=1.12.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-chroma>=0.1.0
chromadb>=0.4.0
pypdf>=3.17.0
tiktoken>=0.5.0
streamlit>=1.28.0
faiss-cpu>=1.7.4
ragas>=0.1.0
datasets>=2.14.0
pandas>=2.0.0